refactor(cli): A.3.2 业务 inline evolve/sample 到 oclif Command
evolve / sample：业务整段 inline 到 src/cli/oclif/commands/{evolve,sample}.ts。
删 src/cli/commands/{evolve,sample}.ts。

runEvolve / runSample 作 module-level helper export from oclif Command file，
方便 cli-exit.test / sample-fix.test 走 in-process import 验证业务（test 改
import path 从 commands/* → oclif/commands/*）。

sample.ts 内 type `Sample` 跟 Command class `Sample` 同名冲突，type import
rename 为 `SampleType`。

yarn lint && build && test 1494 全绿。
